consolidate termslists into one
