+++ src/components/Navbar.svelte
@@ -45,7 +45,7 @@
 
 <nav class="navbar is-black" role="navigation" aria-label="main navigation">
   <div class="navbar-brand">
-    <a class="navbar-item" href="/">
+    <a class="navbar-item" href="#/">
       <img class="logo" src="kumu-logo.png" height="28" alt="levitation logo" />
       <span class="title is-4">InstaQuiz</span>
     </a>
