+++ src/components/courses/AddStudentToCourse.svelte
@@ -1,4 +1,5 @@
 <script>
+  import { push } from 'svelte-spa-router'
   import { mutate } from 'svelte-apollo'
   import { notifications } from '../notifications'
   import { auth } from '../../data/auth'
@@ -31,6 +32,7 @@
       })
       errors = ''
       notifications.add({ text: `Student ${$auth.name} added to ${response.data.addStudentToCourse.name}`, type: 'success' })
+      push('/')
     } catch (error) {
       errors = error
       notifications.add({ text: 'registration failed.', type: 'danger' })
