+++ src/components/terms/Terms.svelte
@@ -1,14 +1,7 @@
 <script>
-  import { TERMS_AND_ALL } from '../../data/queries'
-  import { query } from 'svelte-apollo'
-  import { client } from '../../data/apollo'
-  import TermsList from './TermsList.svelte'
-  import Loading from '../Loading.svelte'
-  import Error from '../Error.svelte'
+  import TermsLoader from './TermsLoader.svelte'
   import { auth } from '../../data/auth'
   import CreateTerm from './CreateTerm.svelte'
-
-  const termsCache = query(client, { query: TERMS_AND_ALL })
 </script>
 
 <svelte:head>
@@ -28,10 +21,4 @@
   <CreateTerm />
 {/if}
 
-{#await $termsCache}
-  <Loading what="terms and courses" />
-{:then result}
-  <TermsList terms={result.data.terms}/>
-{:catch errors}
-  <Error {errors}/>
-{/await}
+<TermsLoader />
